+++ winpr/libwinpr/sspi/sspi_gss.c
@@ -74,7 +74,7 @@ BOOL sspi_gss_wrap_token(SecBuffer* buf, const WinPrAsn1_OID* oid, uint16_t tok_
 	Stream_StaticInit(&s, buf->pvBuffer, len);
 	if (WinPrAsn1EncToStream(enc, &s))
 	{
-		buf->cbBuffer = len;
+		buf->cbBuffer = (UINT32)len;
 		ret = TRUE;
 	}
